### 2024-07-03 #6956 @editorialbot
[PRE REVIEW]: BlueCelluLab: Biologically Detailed Neural Network Experimentation API
**Submitting author:** @anilbey (<a href="http://orcid.org/0000-0003-0317-2556">Mustafa Anıl Tuncel</a>)
**Repository:** https://github.com/BlueBrain/BlueCelluLab
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.6.16
