### 2024-07-03 #6956 @Kevin-Mattheus-Moerman
@anilbey Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps, and to find a handling editor. Before we proceed, please can you have a look at the following points:

Secondly, as you can see I have just added the `waitlisted` label. This is because currently all editors in this domain are preoccupied with other submissions. We will assign an editor as soon as one becomes available.
