### 2024-07-05 #6956 @anilbey
Hi @Kevin-Mattheus-Moerman thanks for mentioning the two points. I addressed the CONTRIBUTING file here https://github.com/BlueBrain/BlueCelluLab/pull/190.
