+++ paper/paper.bib
@@ -58,6 +58,7 @@
 @book{kernighan1984unix,
   title={The UNIX programming environment},
   author={Kernighan, Brian W and Pike, Rob},
+  doi = {10.1002/spe.4380090102},
   volume={270},
   year={1984},
   publisher={Prentice-Hall Englewood Cliffs, NJ}
