+++ paper/paper.bib
@@ -80,6 +80,7 @@
 @book{martin2003agile,
   title={Agile software development: principles, patterns, and practices},
   author={Martin, Robert Cecil},
+  doi={10.1002/pfi.21408},
   year={2003},
   publisher={Prentice Hall PTR}
 }
