+++ paper/paper.bib
@@ -88,6 +88,7 @@
 @article{parnas1972criteria,
   title={On the criteria to be used in decomposing systems into modules},
   author={Parnas, David Lorge},
+  doi={10.1007/978-3-642-59412-0_26},
   journal={Communications of the ACM},
   volume={15},
   number={12},
