### 2024-07-05 #6956 @anilbey
I added the 3/6 of the missing DOIs. To my knowledge the remaining 3 don't have DOIs available.
