### 2024-07-08 1f108a4
add cvode argument to apply_multiple_stimuli
