### 2024-07-08 38a08a7
support custom node set defined in simulation config
