### 2024-07-09 3c7bcae
Add holding current to apply_multiple_stimuli
