### 2024-07-09 47d1931
Correct 'projections' for get_syn_descriptions (#187)
