### 2024-07-16 451a844
replace amp_cv with relative_skew
