+++ tox.ini
@@ -44,7 +44,7 @@ commands =
     ruff check . --select F541,F401 --per-file-ignores="__init__.py:F401"
     pycodestyle {[base]name} --ignore=E501,W504,W503
     pycodestyle tests --ignore=E501,W504,W503,E741
-    mypy . --ignore-missing-imports --disable-error-code=call-overload  # remove once pandas-stubs makes a release after 1.5.3.230321 
+    mypy . --ignore-missing-imports --disable-error-code=call-overload  # remove once pandas-stubs makes a release after 1.5.3.230321
     docformatter --check bluecellulab -r
 
 [testenv:examples]
