### 2024-07-17 adfbb4f
remove double CircuitSimulation (#196)
Co-authored-by: Jaquier Aurélien Tristan <aurelien.jaquier@epfl.ch>
