+++ bluecellulab-bb5-tests/tests/test_ssim_gpfs.py
@@ -1,0 +1,517 @@
+"""Unit tests for SSim"""
+
+# pylint: disable=E1101,W0201,F0401,E0611,W0212
+
+import itertools
+import os
+
+import numpy as np
+import pytest
+
+import bluecellulab
+from bluecellulab import CircuitSimulation
+from bluecellulab.circuit import SynapseProperty
+from bluecellulab.circuit.node_id import create_cell_id
+from bluecellulab.circuit.circuit_access import get_synapse_connection_parameters
+from bluecellulab.circuit.config.sections import Conditions
+
+script_dir = os.path.dirname(__file__)
+
+# BB5 compiled NEURON is using the legacy units!
+bluecellulab.neuron.h.nrnunit_use_legacy(True)
+
+
+def proj_path(n):
+    return f"/gpfs/bbp.cscs.ch/project/proj{n}/"
+
+
+# Example ReNCCv2 sim used in BluePy use cases
+renccv2_bc_1_path = os.path.join(
+    proj_path(1),
+    "simulations/ReNCCv2/k_ca_scan_dense/K5p0/Ca1p25_synreport/",
+    "BlueConfig",
+)
+
+test_single_vesicle_path = os.path.join(
+    proj_path(83), "home/tuncel/bglibpy-tests/single-vesicle-AIS", "BlueConfig"
+)
+
+plasticity_sim_path = os.path.join(
+    proj_path(96), "home/tuncel/simulations/glusynapse-pairsim/p_Ca_2_0/", "BlueConfig"
+)
+
+risetime_sim_path = os.path.join(
+    proj_path(96),
+    "home/tuncel/simulations/glusynapse-pairsim/p_Ca_2_0_randomize_risetime/",
+    "BlueConfig",
+)
+
+no_rand_risetime_sim_path = os.path.join(
+    proj_path(96),
+    "home/tuncel/simulations/glusynapse-pairsim/p_Ca_2_0_randomize_risetime/",
+    "BlueConfigNoRandRise",
+)
+
+
+test_shotnoise_path = os.path.join(
+    proj_path(83), "home/tuncel/bglibpy-tests/shotnoise-sim", "BlueConfig"
+)
+
+test_relative_shotnoise_path = os.path.join(
+    proj_path(83), "home/bolanos/bglibpy-tests/relative_shotnoise", "BlueConfig"
+)
+
+test_relative_shotnoise_conductance_path = os.path.join(
+    proj_path(83),
+    "home/bolanos/bglibpy-tests/relative_shotnoise_conductance",
+    "BlueConfig",
+)
+
+test_ornstein_path = os.path.join(
+    proj_path(83), "home/tuncel/bglibpy-tests/ornstein_uhlenbeck", "BlueConfig"
+)
+
+test_relative_ornstein_path = os.path.join(
+    proj_path(83),
+    "home/tuncel/bglibpy-tests/relative_ornstein_uhlenbeck/",
+    "BlueConfig",
+)
+
+
+@pytest.mark.v5
+class TestSSimBaseClass_full_run:
+    """Class to test SSim with full circuit"""
+
+    def setup_method(self):
+        """Setup"""
+        self.t_start = 0
+        self.t_stop = 500
+        self.record_dt = 0.2
+        self.len_voltage = self.t_stop / self.record_dt
+        self.ssim = CircuitSimulation(renccv2_bc_1_path, record_dt=self.record_dt)
+        assert isinstance(self.ssim, CircuitSimulation)
+
+    def test_run(self):
+        """SSim: Check if a full replay of a simulation run with forwardskip
+        gives the same output trace as on BGQ"""
+        gid = 75936
+        self.ssim.instantiate_gids(
[…]
