+++ bluecellulab-bb5-tests/tests/examples/sim_twocell_empty/run.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+BGLIB_DIR=~/private/src/bbp/lib
+
+$BGLIB_DIR/x86_64/special $BGLIB_DIR/hoclib/init.hoc -NFRAME 1000 -mpi
+cd output
+cp out.dat.original out.dat
