+++ bluecellulab-bb5-tests/.gitlab-ci.yml
@@ -1,4 +1,0 @@
-include:
-   - project: cells/ci
-     file:
-        - /ci/pipelines/cells.bluecellulab-bb5-tests.yml
