+++ bluecellulab-bb5-tests/Makefile
@@ -3,7 +3,7 @@ PIP_INDEX_URL := https://bbpteam.epfl.ch/repository/devpi/bbprelman/dev/+simple
 PIP_EXTRA_INDEX_URL := https://pypi.python.org/simple
 NEURON_MODULE_OPTIONS := '-nogui'
 TESTS_DIR := tests
-MECHANISMS_DIR := tests/mechanisms
+MECHANISMS_DIR := ./../tests/mechanisms/
 PYTEST_OPTS := --numprocesses=auto --timeout=600 -vsx --forked
 
 # Default rule (similar to 'testenv' in tox)
