+++ bluecellulab-bb5-tests/tests/examples/sim_twocell_replay/run.sh
@@ -1,4 +1,19 @@
 #!/usr/bin/env bash
+
+# Copyright 2024 Blue Brain Project / EPFL
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 BGLIB_DIR=~/private/src/bbp/lib
 
 $BGLIB_DIR/x86_64/special $BGLIB_DIR/hoclib/init.hoc -NFRAME 1000 -mpi
