+++ bluecellulab-bb5-tests/tests/test_synapse_parameters.py
@@ -1,3 +1,16 @@
+# Copyright 2024 Blue Brain Project / EPFL
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 """Tests the synapse parameters."""
 
 import os
@@ -23,7 +36,7 @@ proj83_path = "/gpfs/bbp.cscs.ch/project/proj83/"
 def assert_series_almost_equal(series1, series2, rtol=1e-05, atol=1e-08):
     # Check if both series have the same length
     assert len(series1) == len(series2), "The series have different lengths."
-    
+
     # Iterate over each pair of elements
     for i in range(len(series1)):
         # Check for string first
