+++ .compile_mod.sh
@@ -1,12 +1,26 @@
 #!/bin/sh
 
+# Copyright 2012-2024 Blue Brain Project / EPFL
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 set -e
 
 INSTALL_DIR=$1
 MOD_DIR=$2
 
 cd ${INSTALL_DIR}
- 
+
 echo "Building mod files"
 rm -rf x86_64
 nrnivmodl ${MOD_DIR} >nrnivmodl.log 2>&1
