### 2024-07-30 #7026 @finsberg
@mooniean I am done with my review on this software and can recommend it for publication in JOSS. The authors did an excellent job preparing this software for JOSS.
