### 2024-08-01 c5e54a0
CI(dependabot): Bump anothrNick/github-tag-action from 1.69.0 to 1.70.0
Bumps [anothrNick/github-tag-action](https://github.com/anothrnick/github-tag-action) from 1.69.0 to 1.70.0.
- [Release notes](https://github.com/anothrnick/github-tag-action/releases)
- [Commits](https://github.com/anothrnick/github-tag-action/compare/1.69.0...1.70.0)

---
updated-dependencies:
- dependency-name: anothrNick/github-tag-action
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
