+++ .github/workflows/release.yml
@@ -25,7 +25,7 @@ jobs:
           python-version: "3.10"
 
       - name: Fetch the tag, if any. If not tagged, create a patch tag
-        uses: anothrNick/github-tag-action@1.69.0
+        uses: anothrNick/github-tag-action@1.70.0
         if: ${{ !startsWith(github.ref, 'refs/tags/') }}
         id: tag
         env:
