### 2024-08-05 #7026 @ryEllison
Good morning, @mooniean! I have now finished my review of the paper/software. It's a rather nice frontend API to the NEURON simulation engine, facilitating quick neuron-model development. There's certainly an off-the-shelf compatibility during installation for Unix-like OS over Windows. With that said, however, I can make the recommendation for publication in JOSS.
