### 2024-08-05 #7026 @mooniean
This is excellent, thank you both for your reviews!

Before going to the post-review checklist, @anilbey: in the paper, can you add a small explanation for Figure 1? It is not mentioned at any point in the text, it's only for clarification!

After, I'll be happy to advance to the next steps.
