### 2024-08-06 511ea26
add bb5-tests (#197)
* add bb5-tests

* remove gitlab-ci.yml

* remove duplicate mod files for bb5-tests and merge them into the 'tests' folder

* mod files info

* add license header

* update Makefile

* update readme
