+++ bluecellulab-bb5-tests/tests/test_circuit_access.py
@@ -1,0 +1,357 @@
+# Copyright 2024 Blue Brain Project / EPFL
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+"""Unit tests for the circuit_access module."""
+from pathlib import Path
+from unittest.mock import patch
+
+import pandas as pd
+import pytest
+
+from bluecellulab.circuit import BluepyCircuitAccess, CellId
+from bluecellulab.exceptions import BluecellulabError
+from bluecellulab.circuit.config.sections import ConnectionOverrides
+
+from helpers.circuit import blueconfig_append_path
+
+parent_dir = Path(__file__).resolve().parent
+proj55_path = "/gpfs/bbp.cscs.ch/project/proj55/"
+
+test_thalamus_path = (
+    Path(proj55_path)
+    / "tuncel/simulations/release"
+    / "2020-08-06-v2"
+    / "bglibpy-thal-test-with-projections"
+    / "BlueConfig"
+)
+
+
+def test_non_existing_circuit_config():
+    with pytest.raises(FileNotFoundError):
+        BluepyCircuitAccess(
+            str(parent_dir / "examples" / "non_existing_circuit_config")
+        )
+
+
+@pytest.mark.thal
+def test_get_connectomes_dict_with_projections():
+    """Test the retrieval of projection and the local connectomes."""
+    circuit_access = BluepyCircuitAccess(str(test_thalamus_path))
+
+    # empty
+    assert circuit_access._get_connectomes_dict(None).keys() == {""}
+
+    # single projection
+    connectomes_dict = circuit_access._get_connectomes_dict("ML_afferents")
+    assert connectomes_dict.keys() == {"", "ML_afferents"}
+
+    # multiple projections
+    all_connectomes = circuit_access._get_connectomes_dict(
+        ["ML_afferents", "CT_afferents"]
+    )
+    assert all_connectomes.keys() == {"", "ML_afferents", "CT_afferents"}
+
+
+@pytest.mark.thal
+def test_get_all_projection_names():
+    """Test the retrieval of projection and the local connectomes."""
+    circuit_access = BluepyCircuitAccess(str(test_thalamus_path))
+
+    assert set(circuit_access.config.get_all_projection_names()) == {
+        "CT_afferents",
+        "ML_afferents",
+    }
+
+
+@pytest.mark.thal
+def test_get_population_ids():
+    """Test the retrieval of population ids."""
+    circuit_access = BluepyCircuitAccess(str(test_thalamus_path))
+    assert circuit_access.get_population_ids("ML_afferents") == (1, 0)
+    assert circuit_access.get_population_ids("CT_afferents") == (2, 0)
+
+
+class TestCircuitAccess:
+    def setup_method(self):
+        conf_pre_path = parent_dir / "examples" / "sim_twocell_all"
+
+        # make the paths absolute
+        modified_conf = blueconfig_append_path(
+            conf_pre_path / "BlueConfigWithConditions", conf_pre_path
+        )
+        self.circuit_access = BluepyCircuitAccess(modified_conf)
+
+    def test_target_contains_cell(self):
+        target = "Mosaic"
+        cell_id = CellId("", 1)
+        assert self.circuit_access.target_contains_cell(target=target, cell_id=cell_id)
+
[…]
