### 2024-08-07 a513dfe
add explanation for figure1
