### 2024-08-09 b639a8d
add explanation for figure1 (#199)
* add cvode argument to apply_multiple_stimuli (#191)

Co-authored-by: Jaquier Aurélien Tristan <aurelien.jaquier@epfl.ch>

* Feat: support custom node sets defined in simulation config (#192)

* Add holding current to apply_multiple_stimuli (#193)

Co-authored-by: Jaquier Aurélien Tristan <aurelien.jaquier@epfl.ch>

* Correct 'projections' for get_syn_descriptions (#187)

* add contributing guidelines (#190)

* replace amp_cv with relative_skew (#195)

* remove double CircuitSimulation (#196)

Co-authored-by: Jaquier Aurélien Tristan <aurelien.jaquier@epfl.c
[…]
