### 2024-08-09 #7026 @anilbey
Thanks reviewers for taking your time and reviewing BlueCelluLab. @mooniean I just added it the explanation here https://github.com/BlueBrain/BlueCelluLab/pull/199 let me regenerate the pdf.
