### 2024-08-09 #7026 @mooniean
@anilbey I've started the post-review checklist. Can you please follow the tasks in my previous comment?

> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * Double check authors and affiliations (including ORCIDs)
> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * Archive the release on Zenodo/figshare/etc and post the DOI here.
> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * Make sure that the license listed for the archive is the same as the software license.

Furthermore, there seems to be 3 DOIs missing, could you try to address those please?
