### 2024-08-10 b8a92bb
update mypy reference
