### 2024-08-10 e15428e
update mypy citation url
