+++ paper/paper.bib
@@ -107,7 +107,7 @@
 @article{lehtosalo2017mypy,
   title={Mypy-optional static typing for python},
   author={Lehtosalo, Jukka and Rossum, G v and Levkivskyi, Ivan and Sullivan, Michael J and Fisher, David and Price, Greg and Lee, Michael and Seyfer, N and Barton, R and Ilinskiy, S and others},
-  journal={URL: http://mypy-lang. org/[cited 2021-11-30]},
+  journal={URL: https://mypy-lang.org},
   year={2017}
 }
 
