### 2024-08-12 54e216c
update mypy reference (#201)
* update mypy reference

* update mypy citation url
