+++ paper/paper.bib
@@ -104,12 +104,13 @@
   year={2014}
 }
 
-@phdthesis{lehtosalo2015adapting,
-  title={Adapting dynamic object-oriented languages to mixed dynamic and static typing},
-  author={Lehtosalo, Jukka Antero},
-  year={2015},
-  school={University of Cambridge}
+@article{lehtosalo2017mypy,
+  title={Mypy-optional static typing for python},
+  author={Lehtosalo, Jukka and Rossum, G v and Levkivskyi, Ivan and Sullivan, Michael J and Fisher, David and Price, Greg and Lee, Michael and Seyfer, N and Barton, R and Ilinskiy, S and others},
+  journal={URL: https://mypy-lang.org},
+  year={2017}
 }
+
 
 @software{pydantic_v2.6.4,
   author       = {Samuel Colvin and Eric Jolibois and Hasan Ramezani and Adrian Garcia Badaracco and Terrence Dorsey and David Montague and Serge Matveenko and Marcelo Trylesinski and Sydney Runkle and David Hewitt and Alex Hall},
