### 2024-08-12 #7026 @AurelienJaquier
Here is the DOI for the latest release on zenodo: https://doi.org/10.5281/zenodo.13303087
