### 2024-08-12 b710c79
Update readme citation and add CITATION.cff file
