### 2024-08-12 c5666ce
Update readme citation and add CITATION.cff file (#203)
Co-authored-by: Jaquier Aurélien Tristan <aurelien.jaquier@epfl.ch>
