### 2024-08-12 #7026 @anilbey
Hi @mooniean

> Furthermore, there seems to be 3 DOIs missing, could you try to address those please?

This was also mentioned during the pre-review https://github.com/openjournals/joss-reviews/issues/6956#issuecomment-2211381582.

Those 3 references do not have the DOI. If all references require a DOI, can we turn those three into footnotes instead (if the markdown template supports it)?
