### 2024-08-13 #7026 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
