### 2024-08-14 4a794b6
Update copyright year (#204)
* update copyright year

* move bb5-tests into tests
