### 2024-08-15 be0a19e
exclude example-7 in CI to reduce test time
