+++ tox.ini
@@ -57,7 +57,7 @@ deps =
     scipy>=1.11.1
     seaborn>=0.12.2
 commands =
-    pytest --ignore=tests/7-Extra-Simulation/ --numprocesses=auto --forked --nbmake examples # exclude long tests
+    pytest --ignore=examples/7-Extra-Simulation/ --numprocesses=auto --forked --nbmake examples # exclude long tests
 
 [testenv:docs]
 envdir =
