### 2024-08-15 fe4baed
minor improvement to sonata-network example (#205)
* improve sonata example

* exclude example-7 in CI to reduce test time
