+++ tox.ini
@@ -27,7 +27,7 @@ setenv =
 commands =
     make clean
     ./.compile_mod.sh . tests/mechanisms
-    pytest {[testenv]coverage_options} tests --ignore=tests/bluecellulab-bb5-tests/ --numprocesses=auto --timeout=600 -vsx --forked
+    pytest {[testenv]coverage_options} tests --ignore=tests/bluecellulab-bb5-tests/ --numprocesses=auto --timeout=600 -vsx --forked # exclude bb5-tests for now
     coverage report --show-missing
     coverage xml
 
@@ -57,7 +57,7 @@ deps =
     scipy>=1.11.1
     seaborn>=0.12.2
 commands =
-    pytest --numprocesses=auto --forked --nbmake examples
+    pytest --ignore=examples/7-Extra-Simulation/ --numprocesses=auto --forked --nbmake examples # exclude long tests
 
 [testenv:docs]
 envdir =
