### 2024-08-16 #7026 @Kevin-Mattheus-Moerman
- [ ] I see you have additional released created in the last couple of days, should we here update to `2.6.32` or is `2.6.28` still accurate for this review?
- [ ] The following is a recommendation, not a requirement. I would remove the text at the top of you LICENSE (which mentions the 3rd party/dependency licenses), since  you already mention this in your README. This way you LICENSE file can fully match the normal Apache 2.0 text and GitHub (and our systems too) would easily recognise it then as such.
