### 2024-08-17 #7026 @anilbey
Hi @Kevin-Mattheus-Moerman, yes please update the version to be 2.6.32 since it contains changes related to the copyright and citation.
