### 2024-08-17 BlueBrain/BlueCelluLab#206 @anilbey
Keep main Apache software license intact and mention the modelDB and Zenodo in the examples directory
This is mentioned by the AEiC of JOSS
https://github.com/openjournals/joss-reviews/issues/7026#issuecomment-2293028326

BlueCelluLab **does not depend on** any of those licensed morphology or MOD files to be installed and run.

The MOD and morphologies are not even part of the package resources.
https://github.com/BlueBrain/BlueCelluLab/blob/fe4baed5db1cd3df85012e22e3625fb0646c5b5b/MANIFEST.in

Only the example directories and test directories show its applicability using those files.

So I suggest this commit can be reverted so that the software will have the original Apache license text
https://github.com/BlueBrain/BlueCelluLab/commit/baebfcb8b58f61b61fd5ec39be885672ce55ae4e

And then 2 LICENSE files can be created inside the /examples and /tests to acknowledge where the test and examples data/code (morph and MOD) comes from.
