### 2024-08-17 #7026 @anilbey
>  The following is a recommendation, not a requirement. I would remove the text at the top of you LICENSE (which mentions the 3rd party/dependency licenses), since you already mention this in your README. This way you LICENSE file can fully match the normal Apache 2.0 text and GitHub (and our systems too) would easily recognise it then as such.

I totally agree. Especially because those 3rd party code/data are not part of the package. Those are only used in the example directory to demonstrate some real world use cases. I created this issue here to address it. https://github.com/BlueBrain/BlueCelluLab/issues/206
